# Improve performance of hex string parsing in SKColor