**Description of Change**

Improve performance of hex string parsing in SKColor using spans when available and with fewer allocated strings when spans are not supported.

### Benchmark
- **ParseOld** is the current implementation for parsing hex strings to SKColor.
- **ParseImproved** is the improved parsing logic for platforms/frameworks that don't support spans
- **ParseSpan** is the new parsing logic using spans to avoid allocating strings
```
|        Method |                  Job |              Runtime |          Mean |      Error |     StdDev | Ratio | RatioSD |  Gen 0 | Allocated |
|-------------- |--------------------- |--------------------- |--------------:|-----------:|-----------:|------:|--------:|-------:|----------:|
|      ParseOld |             .NET 6.0 |             .NET 6.0 | 1,663.7469 ns | 32.9390 ns | 43.9727 ns |  1.00 |    0.00 | 1.8501 |   3,872 B |
| ParseImproved |             .NET 6.0 |             .NET 6.0 |   833.2502 ns | 16.1750 ns | 15.1301 ns |  0.50 |    0.02 | 0.5655 |   1,184 B |
|     ParseSpan |             .NET 6.0 |             .NET 6.0 |   507.7845 ns |  3.4736 ns |  3.2492 ns |  0.31 |    0.01 |      - |         - |
|               |                      |                      |               |            |            |       |         |        |           |
|      ParseOld |        .NET Core 3.1 |        .NET Core 3.1 | 1,981.3356 ns | 18.0860 ns | 16.0328 ns |  1.00 |    0.00 | 1.8501 |   3,872 B |
| ParseImproved |        .NET Core 3.1 |        .NET Core 3.1 | 1,172.5769 ns | 23.1321 ns | 31.6635 ns |  0.60 |    0.02 | 0.5646 |   1,184 B |
|     ParseSpan |        .NET Core 3.1 |        .NET Core 3.1 |   786.5793 ns |  2.1876 ns |  1.8267 ns |  0.40 |    0.00 |      - |         - |
|               |                      |                      |               |            |            |       |         |        |           |
|      ParseOld | .NET Framework 4.7.2 | .NET Framework 4.7.2 | 4,420.4635 ns | 26.6690 ns | 24.9462 ns | 1.000 |    0.00 | 2.1973 |   4,622 B |
| ParseImproved | .NET Framework 4.7.2 | .NET Framework 4.7.2 | 3,542.1632 ns | 23.6055 ns | 22.0806 ns | 0.801 |    0.01 | 0.6943 |   1,460 B |
```

Existing tests for parsing hex strings cover the changes of this PR.

**API Changes**

None.

**Behavioral Changes**

None.

**Required skia PR**

None.

**PR Checklist**

- [x] Has tests (if omitted, state reason in description)
- [x] Rebased on top of main at time of PR
- [ ] Merged related skia PRs
- [x] Changes adhere to coding standard
- [ ] Updated documentation