@jwikberg 2023-06-18
> I see some parsing tests are failing: https://dev.azure.com/xamarin/public/_build/results?buildId=88527&view=ms.vss-test-web.build-test-results-tab

This should be fixed now